Hoist the rest of css/matching.rs into style/.
