+++ components/layout/lib.rs
@@ -96,7 +96,3 @@ mod table_wrapper;
 mod text;
 mod traversal;
 mod wrapper;
-
-mod css {
-    pub mod matching;
-}
