+++ components/style/Cargo.toml
@@ -38,5 +38,6 @@ string_cache = "0.2"
 euclid = {version = "0.4", features = ["plugins"]}
 serde = "0.6"
 serde_macros = "0.6"
+time = "0.1"
 url = "0.5.2"
