+++ components/style/lib.rs
@@ -40,6 +40,7 @@ extern crate smallvec;
 #[macro_use(atom, ns)] extern crate string_cache;
 #[macro_use]
 extern crate style_traits;
+extern crate time;
 extern crate url;
 extern crate util;
